% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path-to-file.R
\name{path_to_file}
\alias{path_to_file}
\title{Get file path to `afribats.csv` and `afribats_sf.gpkg` files}
\source{
This function is adapted from `readxl::readxl_example()`.
}
\usage{
path_to_file(path = NULL)
}
\arguments{
\item{path}{Name of file in quotes with extension;
`"afribats.csv"` and `"afribats_sf.gpkg"` will work.
If `NULL`, the example files will be listed.}
}
\description{
afribat comes bundled with one csv and one gpkg file in its `inst/extdata`
directory. This function make them easy to access.
}
\details{
The data in each file is the almost the same. In gpkg file, all missing positions coordinates were removed.
}
\examples{
# load libraries
library(sf)
library(readr)

path_to_file("afribats.csv")
head(read_csv(path_to_file("afribats.csv")))
head(read_sf(path_to_file("afribats_sf.gpkg")))

}
