% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity-indices.R
\name{shannon_weiner}
\alias{shannon_weiner}
\title{Calculate Shannon-Weiner Diversity Index}
\usage{
shannon_weiner(x)
}
\arguments{
\item{x}{A character vector of species names. Each element represents an individual of a species.}
}
\value{
A numeric value representing the Shannon-Weiner diversity index. Returns \code{NA} if the input
vector is empty or contains fewer than two unique species.
}
\description{
This function calculates the Shannon-Weiner diversity index, a measure of species diversity
based on species relative abundance.
}
\note{
The Shannon-Weiner index is undefined for an empty vector or vectors with only one unique species.
For such cases, the function will return \code{NA} or 0 as appropriate.
}
\examples{
# Example 1: Calculate Shannon-Weiner index for a dataset
x <- c("jonesi","gambiensis","hispida","hispida","thebaica","helvum","helvum","gambianus","gambianus","pusillus","monstrosus","aegyptiacus")
shannon_weiner(x)

# Example 2: Handle edge cases
shannon_weiner(c("jonesi")) # Returns 0
shannon_weiner(character(0)) # Returns NA with a warning

}
